+++ packages/ui/src/components/autocomplete/autocomplete.tsx
@@ -27,7 +27,7 @@ function AutocompleteInput({
   return (
     <AutocompletePrimitive.InputGroup
       data-slot="autocomplete-input-group"
-      className="relative w-full text-foreground has-disabled:opacity-50"
+      className="relative w-full text-foreground"
     >
       {startAddon && (
         <div
@@ -41,7 +41,7 @@ function AutocompleteInput({
       <AutocompletePrimitive.Input
         data-slot="autocomplete-input"
         className={cn(
-          "flex h-9 w-full min-w-0 rounded-lg border border-input bg-transparent px-2.5 py-1 text-base transition-colors outline-none placeholder:text-muted-foreground focus-visible:border-ring focus-visible:ring-3 focus-visible:ring-ring/50 disabled:pointer-events-none disabled:cursor-not-allowed disabled:opacity-50 aria-invalid:border-destructive md:text-sm dark:bg-input/30",
+          "flex h-9 w-full min-w-0 rounded-lg border border-input bg-transparent px-2.5 py-1 text-base transition-colors outline-none placeholder:text-muted-foreground focus-visible:border-ring focus-visible:ring-3 focus-visible:ring-ring/50 disabled:pointer-events-none disabled:cursor-not-allowed disabled:bg-input/50 disabled:opacity-50 aria-invalid:border-destructive md:text-sm dark:bg-input/30 dark:disabled:bg-input/80",
           startAddon && "ps-8",
           (showTrigger || showClear) && "pe-8",
           className,